+++ src/IdentityAccess/Contract/Actor.php
@@ -1,0 +1,13 @@
+<?php
+
+/*
+ * SPDX-FileCopyrightText: 2026 The Kadupul project and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+namespace Kadupul\IdentityAccess\Contract;
+
+final readonly class Actor
+{
+    public function __construct(public int $id, public string $username) {}
+}