+++ tests/security/build_architectural_helper_report.sh
@@ -13,6 +13,7 @@ MODE="${1:---summary}"
 
 EXCLUDE=(
 	--glob '!include/vendor/**'
+	--glob '!include/fa/**'
 	--glob '!node_modules/**'
 	--glob '!var/**'
 	--glob '!locales/**'